Allow a step to not have any code archive while cleaning

+++ api/src/ContinuousPipe/Builder/Aggregate/BuildStep/BuildStep.php
@@ -155,7 +155,9 @@ class BuildStep
 
     public function cleanUp()
     {
-        $this->codeArchive->delete();
+        if (null !== $this->codeArchive) {
+            $this->codeArchive->delete();
+        }
 
         foreach ($this->createdArchives as $archive) {
             $archive->delete();
